ref(compiler/ssa)[mnt]: move traits into `ssa` related crate
