+++ compiler/tidec_codegen_llvm/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 # tidy-alphabetical-start
 inkwell = { version = "0.5.0", features = ["llvm18-0"] } # inkwell = { git = "https://github.com/stevefan1999-personal/inkwell", rev = "0732f8dcb7b2b7f8edc25895d6dbd37ba439672c", features = [ "llvm19-1" ] }
 tidec_abi = { path = "../tidec_abi" }
+tidec_codegen_ssa = { path = "../tidec_codegen_ssa" }
 tidec_lir = { path = "../tidec_lir" }
 tidec_utils = { path = "../tidec_utils" }
 tracing = "0.1.41"
